# PERF: Queries on `user_emails.email` column not using the index.