--- a/db/migrate/20180316072756_alter_index_email_on_user_emails.rb
+++ b/db/migrate/20180316072756_alter_index_email_on_user_emails.rb
@@ -1,0 +1,11 @@
+class AlterIndexEmailOnUserEmails < ActiveRecord::Migration[5.1]
+  def up
+    execute("DROP INDEX index_user_emails_on_email")
+    execute "CREATE UNIQUE INDEX index_user_emails_on_email ON user_emails (email);"
+  end
+
+  def down
+    execute("DROP INDEX index_user_emails_on_email")
+    execute "CREATE UNIQUE INDEX index_user_emails_on_email ON user_emails (lower(email));"
+  end
+end