@tgxworld 2018-03-19
Hmm sorry this wouldn't work because our index has to be case sensitive. Will fix the queries instead.